updated scoring

+++ src/main/java/cecs429/weight/Strategy.java
@@ -1,9 +1,8 @@
 package cecs429.weight;
 
-import cecs429.index.DiskPositionalIndex;
 
 public interface Strategy {
-    public Float getWdt(Integer tftd, double tokensInDoc, double aveTokensInCorpus, double aveTftd);
-    public Float getWqt(Integer corpusSize, Integer dft);
+    public Float getWdt(double tftd, double tokensInDoc, double aveTokensInCorpus, double aveTftd);
+    public Float getWqt(double corpusSize, double dft);
     public Float getLd(double docWeight, double byteSize);
 }
